[REFACTOR] Manager Interface분리(#139)
