+++ LionHeart-iOS/LionHeart-iOS/Network/Manager/MangerInterface/BookmarkManager.swift
@@ -1,0 +1,13 @@
+//
+//  BookmarkManager.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol BookmarkManager {
+    func getBookmark() async throws -> BookmarkAppData
+    func postBookmark(model: BookmarkRequest) async throws
+}
