+++ LionHeart-iOS/LionHeart-iOS/Network/Manager/MangerInterface/ChallengeManager.swift
@@ -1,0 +1,12 @@
+//
+//  ChallengeManager.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol ChallengeManager {
+    func inquireChallengeInfo() async throws -> ChallengeData
+}
