+++ LionHeart-iOS/LionHeart-iOS/Network/Manager/MangerInterface/LoginManager.swift
@@ -1,0 +1,12 @@
+//
+//  LoginManager.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol LoginManager {
+    func login(type: LoginType, kakaoToken: String) async throws
+}
