+++ LionHeart-iOS/LionHeart-iOS/Network/Manager/MangerInterface/SplashManager.swift
@@ -1,0 +1,13 @@
+//
+//  SplashManager.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol SplashManager {
+    func reissueToken(token: Token) async throws -> Token?
+    func logout(token: UserDefaultToken) async throws
+}
