+++ LionHeart-iOS/LionHeart-iOS/Network/Manager/MangerInterface/TodayManager.swift
@@ -1,0 +1,12 @@
+//
+//  TodayManager.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/14.
+//
+
+import Foundation
+
+protocol TodayManager {
+    func inquiryTodayArticle() async throws -> TodayArticle
+}
